draw geometry

+++ src/WebGLResources/indexBuffer.ts
@@ -7,7 +7,10 @@ export class IndexBuffer {
     }
 
     public glBuffer: WebGLBuffer | null;
-    public indices: Int32Array | null;
+    /**
+     * WebGL can only use 8 or 16 bit indices? no 32 bit support?
+     */
+    public indices: Int16Array | null;
 
     public release() {
         if (this.glBuffer) {
